# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Backtests": {
        "properties": {
          "backtests": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Backtests",
            "type": "array"
          },
          "banner": {
            "$ref": "#/components/schemas/Banner"
          },
          "research": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Research",
            "type": "array"
          }
        },
        "required": [
          "banner",
          "backtests",
          "research"
        ],
        "title": "Backtests",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Operator token (AQ_API_TOKEN)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Operator views and an allowlisted control plane for the paper/shadow trading platform: the kill switch, queued worker jobs (data, backtests, research, read-only broker verification), CSV uploads, strategy governance up to shadow, runtime settings and the shadow/paper scheduler switch. It cannot place orders, enable live trading, grant LIVE_APPROVED or change deployment secrets. Paper/simulated results and backtests are hypothetical and are not a prediction of future returns.",
    "title": "Adaptive Quant operator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/backtests": {
      "get": {
        "operationId": "backtests_api_v1_backtests_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Backtests"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Backtests",
        "tags": [
          "pages"
        ]
      }
    },
...
    "/api/v1/configuration": {
      "get": {
        "description": "The reviewed YAML configuration, redacted (runtime changes: ``/settings``).",
        "operationId": "configuration_api_v1_configuration_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Configuration"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Configuration",
        "tags": [
          "pages"
        ]
      }
    },
    "/api/v1/cycles": {
      "get": {
        "operationId": "cycles_api_v1_cycles_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cycles",
        "tags": [
          "pages"
        ]
      }
    },
    "/api/v1/cycles/{cycle_id}/explain": {
      "get": {
        "description": "The full stored decision chain of one trading cycle.",
        "operationId": "explain_api_v1_cycles__cycle_id__explain_get",
        "parameters": [
          {
            "in": "path",
            "name": "cycle_id",
            "required": true,
            "schema": {
              "title": "Cycle Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Explain Api V1 Cycles  Cycle Id  Explain Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Explain",
        "tags": [
          "pages"
        ]
      }
    },
...
    "/api/v1/executions": {
      "get": {
        "operationId": "executions_api_v1_executions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Executions",
        "tags": [
          "pages"
        ]
      }
    },
...
    "/api/v1/kill-switch": {
      "get": {
        "operationId": "kill_switch_status_api_v1_kill_switch_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KillSwitchView"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Kill Switch Status",
        "tags": [
          "kill switch"
        ]
      }
    },
    "/api/v1/kill-switch/engage": {
      "post": {
        "description": "Stop new risk-increasing orders. Requires ``confirm`` = \"STOP AUTOMATED TRADING\".\n\nNever rate limited: stopping must always be possible.",
        "operationId": "engage_api_v1_kill_switch_engage_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KillSwitchView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Engage",
        "tags": [
          "kill switch"
        ]
      }
    },
    "/api/v1/kill-switch/release": {
      "post": {
        "description": "Re-enable automated trading.\n\nRequires ``confirm`` = \"RE-ENABLE TRADING\" and a named operator.",
        "operationId": "release_api_v1_kill_switch_release_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KillSwitchView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Release",
        "tags": [
          "kill switch"
        ]
      }
    },
    "/api/v1/orders": {
      "get": {
        "operationId": "orders_api_v1_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "cycle_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cycle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Orders",
        "tags": [
          "pages"
        ]
      }
    },
    "/api/v1/overview": {
      "get": {
        "operationId": "overview_api_v1_overview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Overview",
        "tags": [
          "pages"
        ]
      }
    },
...
    "/api/v1/portfolio": {
      "get": {
        "operationId": "portfolio_api_v1_portfolio_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Portfolio"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Portfolio",
        "tags": [
          "pages"
        ]
      }
    },
...
  }
}